+++ src/main/java/org/example/lab3/controllers/ProductController.java
@@ -5,6 +5,8 @@ import org.example.lab3.enums.KitType;
 import org.example.lab3.model.*;
 import org.example.lab3.repos.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -21,6 +23,10 @@ public class ProductController {
     private WheelsRepo wheelsRepo;
     @Autowired
     private BodyKitRepo bodyKitRepo;
+    @Autowired
+    private WarehouseRepo warehouseRepo;
+    @Autowired
+    private CartRepo cartRepo;
 
 
     @GetMapping(value = "/getAllProducts")
@@ -165,9 +171,30 @@ public class ProductController {
 
 
     @DeleteMapping(value = "/deleteProduct/{id}")
-    public @ResponseBody Optional<Product> deleteProduct(@PathVariable(name = "id") int id) {
-        productRepo.deleteById(id);
-        return productRepo.findById(id);
+    public ResponseEntity<String> deleteProduct(@PathVariable(name = "id") int id) {
+        Optional<Product> productOptional = productRepo.findById(id);
+
+        if (productOptional.isPresent()) {
+            Product product = productOptional.get();
+
+            if (product.getCart() != null) {
+                Cart cart = product.getCart();
+                cart.getItemsToBuy().remove(product);
+                cartRepo.save(cart);
+            }
+
+            if (product.getWarehouse() != null) {
+                Warehouse warehouse = product.getWarehouse();
+                warehouse.getStock().remove(product);
+                warehouseRepo.save(warehouse);
+            }
+
+            productRepo.deleteById(id);
+
+            return ResponseEntity.ok().body("Product with ID " + id + " has been deleted successfully.");
+        } else {
+            return new ResponseEntity<>("Product with ID " + id + " not found.", HttpStatus.NOT_FOUND);
+        }
     }
 
 }
